2020.08.03 20:00 Code Operation Debug #30
TODO :
1) Init FIFO - Done
2) VLCGroup() - Done
3) UpdateHistoryElement - Done
4) addbits, getbits Problem - Done
5) Flatness Test Part Out of Bound Error - Done
6) Final Line Processing - Done
7) FIFO Overflow Problem - Done
8) SamplePredict indexing Problem - Done
9) Syntax Edited - Done
10) Shifter Structure Changed - Done
11) When is Shifter Fullness increased ?? - Done
12) AttributeError: 'numpy.int32' object has no attribute 'bit_length' - Done
13) DSC Write Index Problem - Done
14) Pixel Value Debug ... Priority!!!
15) Output Buffer Write Problem... ONLY WRITES ZERO VALUE!! - Done
16) stQp, 'd' in putbit Value Type Error - Done
17) Line Buffer Value Initialize - Done
18) Zero Padding Position Problems - Done
19) RateControl Fullness and pixelCount Problem ... ing
20) Function Call Test...
    MaxResidualSize
    PopulateOrigLine
    UseICHistory - Done
21) Fine ALL not correct variables... (vlc_var.ichSelected -> ich_var.ichSelected...)
22) Pixel Component Order Problems - Done
23) Predicted Value and Original Pixel Value Compare and Debug - Partially Done
24) BlockPrediction Problem - Done

+++ init_enc_params.py
@@ -39,7 +39,8 @@ class initPredVariables:
         self.max_size = np.zeros(defines.MAX_UNITS_PER_GROUP, ).astype(np.int32)
         self.quantizedResidualSize = np.zeros((defines.MAX_UNITS_PER_GROUP, defines.SAMPLES_PER_UNIT)).astype(np.int32)
         self.bpCount = 0 ## 2020.07.27 added
-        self.prevLinePred = np.zeros(int((dsc_const.sliceWidth + defines.PRED_BLK_SIZE - 1) / defines.PRED_BLK_SIZE), ) ## 2020.07.27 added
+        self.prevLinePred = np.zeros(int((dsc_const.sliceWidth + defines.PRED_BLK_SIZE - 1)
+                                         / defines.PRED_BLK_SIZE), ).astype(np.int32) ## 2020.07.27 added
 
 
 class initFlatVariables:
